Holes, QuadTree merge, & Colinear edges
Fixed an issue relating to holes:
Using a paths (array) object, ClipperLib gives you parent and holes in
chaotic order. Switched to a PolyTree and ExPolygon to get a list of
polygons and their holes.

Fixed an issue with new polygons causing irrelevant polygons still in
the quadtree to be merged out of existence. Fixed by checking an index
in polygons and checking for a quadtree id and dropping if necessary.

Fixed an issue relating to colienar edges:
Just 'cause edges are colinear does not mean they are shared. The
continue statement in Base.mergeInPolygon fixes this.