Transience Hell
Meshes can border each other without sharing an edge. This is through a
transient edge. That is, for one of the polygon's, the other's edge is a
sub-segment of it's own relevant edge. There are particular cases that
are difficult to handle. I am pushing this in order to branch with the
goal of containing it all within the EdgeSet class instead. Will also
make Edges cleaner.