MVis
MVis is an attempt to recreate NavMesh using an annotated mesh instead
of bases. This means navigating objects of different radi will use the
same map. The map will update immediately after adding objects rather
than when a path is requested. Annotations will work on the principle
that if a path between two points is possible at n radius, it is also
possible at [0, n] radius.

Tessellation will not be done. Instead new polygons will search for
internal intersections to produce greedy convex polygons. The merge
algorithm will still perform. Deletions for holes are theoretically
simple by checking for visible points from 1 or 2 of the hole's vertices
(if it is an enclave).

Because polygon offsetting is not required, the use of Clipper may
possibly be dropped, depending on how complex deletions and additions
are. May otherwise switch to a simpler, less buggy, and more active
clipping library since only unions and differences need to be supported.